Reorganized CrawlCtrl unit tests
