--- test/CrawlCtrl.UnitTests/LineTypes/DirectiveTestData.cs
@@ -1,29 +1,0 @@
-namespace CrawlCtrl.UnitTests.LineTypes;
-
-internal static class DirectiveTestData
-{
-    public static IEnumerable<object[]> OriginalDirectiveTestData =>
-        new List<object[]>
-        {
-            new object[] { "Directive", "Directive" },
-            new object[] { " Directive", " Directive" },
-            new object[] { "Directive ", "Directive " },
-            new object[] { " Directive ", " Directive " }
-        };
-    
-    public static IEnumerable<object[]> TrimmedDirectiveTestData =>
-        new List<object[]>
-        {
-            new object[] { "Directive", "Directive" },
-            new object[] { " Directive", "Directive" },
-            new object[] { "Directive ", "Directive" },
-            new object[] { " Directive ", "Directive" }
-        };
-    
-    public static IEnumerable<object[]> EmptyDirectiveTestData =>
-        new List<object[]>
-        {
-            new object[] { "" },
-            new object[] { "  " }
-        };
-}
