+++ test/CrawlCtrl.UnitTests/TestData/DirectiveTestData.cs
@@ -1,0 +1,29 @@
+namespace CrawlCtrl.UnitTests.TestData;
+
+internal static class DirectiveTestData
+{
+    public static IEnumerable<object[]> OriginalDirectiveTestData =>
+        new List<object[]>
+        {
+            new object[] { "Directive", "Directive" },
+            new object[] { " Directive", " Directive" },
+            new object[] { "Directive ", "Directive " },
+            new object[] { " Directive ", " Directive " }
+        };
+    
+    public static IEnumerable<object[]> TrimmedDirectiveTestData =>
+        new List<object[]>
+        {
+            new object[] { "Directive", "Directive" },
+            new object[] { " Directive", "Directive" },
+            new object[] { "Directive ", "Directive" },
+            new object[] { " Directive ", "Directive" }
+        };
+    
+    public static IEnumerable<object[]> EmptyDirectiveTestData =>
+        new List<object[]>
+        {
+            new object[] { "" },
+            new object[] { "  " }
+        };
+}
